+++ package.json
@@ -34,10 +34,7 @@
     "@ng-matero/extensions": "15.4.2",
     "@ngneat/input-mask": "6.0.0",
     "@ngneat/until-destroy": "9.2.2",
-    "@s-libs/js-core": "15.0.0",
-    "@s-libs/micro-dash": "15.0.0",
-    "@s-libs/ng-core": "15.0.0",
-    "@s-libs/rxjs-core": "15.0.0",
+    "@s-libs/ng-core": "^15.0.0",
     "@vality/deanonimus-proto": "2.0.1-2a3d5ad.0",
     "@vality/domain-proto": "2.0.1-bfedcb9.0",
     "@vality/dominant-cache-proto": "2.0.1-99f38c9.0",
